update mesocycle, fixed some bugs meaning that saving was not working properly

+++ TrackMyMacros.App4/ViewModels/Set/UpdateSetViewModel.cs
@@ -12,7 +12,7 @@ namespace TrackMyMacros.App4.ViewModels.Set
         //     ExerciseId = exerciseId;
         // }
         public int Reps { get; init; }
-        public int Weight { get; init; }
+        public double Weight { get; init; }
         public int TargetReps { get; init; }
         public double TargetWeight { get; init; }
         public Guid ExerciseId { get; init; }
